Refactor use roe, remove usdtRoe
