+++ client/src/components/FioDomainStatusChange/index.ts
@@ -2,7 +2,8 @@ import { connect } from 'react-redux';
 import { createStructuredSelector } from 'reselect';
 
 import apis from '../../api';
-import { compose, getElementByFioName, setFees } from '../../utils';
+import { compose, getElementByFioName } from '../../utils';
+import { setFees } from '../../util/prices';
 
 import {
   refreshBalance,
@@ -38,23 +39,27 @@ const reduxConnect = connect(
       const { transactionResult } = state.fio;
       const result = transactionResult[SET_VISIBILITY_REQUEST];
       if (result && result.fee_collected) {
-        const { prices } = state.registrations;
+        const { roe } = state.registrations;
         const feeCollected = result.fee_collected;
         return {
           feeCollected: {
             nativeAmount: feeCollected,
             costFio: apis.fio.sufToAmount(feeCollected),
-            costUsdc: apis.fio.convert(feeCollected, prices.usdtRoe),
+            costUsdc: apis.fio.convert(feeCollected, roe),
           },
         };
       }
 
       return result;
     },
-    feePrice: (state: ReduxState, ownProps: ContainerOwnProps & any) => {
+    feePrice: (state: ReduxState) => {
       const { fees } = state.fio;
-      const { prices } = state.registrations;
-      return setFees(fees[apis.fio.actionEndPoints.setFioDomainPublic], prices);
+      const { prices, roe } = state.registrations;
+      return setFees(
+        fees[apis.fio.actionEndPoints.setFioDomainPublic],
+        prices,
+        roe,
+      );
     },
     domainStatus: (state: ReduxState, ownProps: ContainerOwnProps & any) => {
       // todo: set types for state & fix ownProps type
