+++ client/src/components/FioNameTransfer/index.ts
@@ -2,7 +2,8 @@ import { connect } from 'react-redux';
 import { createStructuredSelector } from 'reselect';
 
 import apis from '../../api';
-import { setFees, hasFioAddressDelimiter } from '../../utils';
+import { hasFioAddressDelimiter } from '../../utils';
+import { setFees } from '../../util/prices';
 
 import {
   refreshBalance,
@@ -35,13 +36,13 @@ const reduxConnect = connect(
       const { transactionResult } = state.fio;
       const result = transactionResult[TRANSFER_REQUEST];
       if (result && result.fee_collected) {
-        const { prices } = state.registrations;
+        const { roe } = state.registrations;
         const feeCollected = result.fee_collected;
         return {
           feeCollected: {
             nativeAmount: feeCollected,
             costFio: apis.fio.sufToAmount(feeCollected),
-            costUsdc: apis.fio.convert(feeCollected, prices.usdtRoe),
+            costUsdc: apis.fio.convert(feeCollected, roe),
           },
           newOwnerKey: result.newOwnerKey,
         };
@@ -51,11 +52,11 @@ const reduxConnect = connect(
     },
     feePrice: (state: ReduxState, ownProps: ContainerOwnProps & any) => {
       const { fees } = state.fio;
-      const { prices } = state.registrations;
+      const { prices, roe } = state.registrations;
       const feeEndPoint = hasFioAddressDelimiter(ownProps.name)
         ? apis.fio.actionEndPoints.transferFioAddress
         : apis.fio.actionEndPoints.transferFioDomain;
-      return setFees(fees[feeEndPoint], prices);
+      return setFees(fees[feeEndPoint], prices, roe);
     },
     walletPublicKey,
     currentWallet,
