+++ client/src/redux/registrations/reducer.ts
@@ -8,7 +8,6 @@ const PRICES_DEFAULT: Prices = {
   fio: { address: 0, domain: 0 },
   fioNative: { address: 0, domain: 0 },
   usdt: { address: 0, domain: 0 },
-  usdtRoe: null,
 };
 
 export default combineReducers({
