+++ client/src/util/prices.ts
@@ -1,0 +1,17 @@
+import { FeePrice, Prices } from '../types';
+import apis from '../api';
+
+export const setFees = (nativeFee: number, prices: Prices, roe: number) => {
+  const fee: FeePrice = {
+    nativeFio: null,
+    costFio: null,
+    costUsdc: null,
+  };
+  fee.nativeFio = nativeFee;
+  fee.costFio = apis.fio.sufToAmount(fee.nativeFio);
+  if (fee.nativeFio && roe) {
+    fee.costUsdc = apis.fio.convert(fee.nativeFio, roe);
+  }
+
+  return fee;
+};
