+++ client/src/utils.ts
@@ -9,7 +9,6 @@ import {
   FioNameItemProps,
   RegistrationResult,
   WalletKeysObj,
-  FeePrice,
 } from './types';
 
 const FIO_DASH_USERNAME_DELIMITER = `.fio.dash.${process.env
@@ -413,21 +412,6 @@ export const getElementByFioName = ({
       )) ||
     {}
   );
-};
-
-export const setFees = (nativeFee: number, prices: Prices) => {
-  const fee: FeePrice = {
-    nativeFio: null,
-    costFio: null,
-    costUsdc: null,
-  };
-  fee.nativeFio = nativeFee;
-  fee.costFio = apis.fio.sufToAmount(fee.nativeFio);
-  if (fee.nativeFio && prices.usdtRoe) {
-    fee.costUsdc = apis.fio.convert(fee.nativeFio, prices.usdtRoe);
-  }
-
-  return fee;
 };
 
 export const putParamsToUrl = (
